Refactor: import, export 구조 수정
