--- src/components/KakaoMap/index.ts
@@ -1,4 +1,0 @@
-import KakaoMap from "./KakaoMap.vue";
-
-export * from "./type";
-export { KakaoMap };
